backend/pyproject.toml
[project]
name = "rag-platform"
version = "1.0.0"
description = "Multi-Tenant RAG Platform - Production-grade document Q&A system"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.3.0",
    "llama-index>=0.12.0",
    "rank-bm25>=0.2.2",
    "cohere>=5.11.0",
    # "ragas>=0.2.0",  # evaluation only - requires C compiler for scikit-network
    "PyMuPDF>=1.25.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.7",
    "argon2-cffi>=23.1.0",
    "structlog>=24.4.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.12",
    "tiktoken>=0.8.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "locust>=2.32.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html:htmlcov --cov-fail-under=40"

[tool.coverage.run]
source = ["app"]
omit = [
    "app/db/migrations/*",
    "app/main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "pass",
    "\\.\\.\\.",
]
show_missing = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
